+++ changelog.md
@@ -4,6 +4,7 @@
 - [FIX][COR] Analytics domain, user agent and multiple sent show event
 - [FIX][IOS] Add templates and some public modifiers
 - [FEA][AND] Change lazy column into lazy grid in recipes list
+- [BRK][FIX][AND] Fix recipes page lazy loading by upgrading compose (1.3.1) and android SDK (33)
 
 ## 3.0.1
 - [FIX][AND] Fix loader on prefrences screen